Tidy up monad instances
